requirements.txt
aws-cdk-lib==2.x.x
constructs>=10.0.0,<11.0.0

